requirements.lite.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis & Celery
redis==5.0.1
celery==5.3.6

# HTTP Clients
httpx==0.26.0
openai==1.10.0

# ML & NLP (Commented out for Lite build)
# sentence-transformers==2.3.1
# torch==2.1.2
# transformers==4.37.0

# Document Processing
pypdf==3.17.4
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.2.0
opencv-python-headless==4.9.0.80

# Utilities
python-dotenv==1.0.0
psutil==5.9.7

# Testing
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
ruff==0.1.14
black==23.12.1
mypy==1.8.0
